Thuc_hanh_Buoi2/requirements.txt
Flask==3.0.0
flask-cors==3.0.10
PyJWT==2.6.0